Add fallback colors to fish highlighting
On 8/16 color terminals, like the Linux console, the color interpolation
on the 24 bit colors in this colorscheme lead to undesirable results,
like the prefix of a valid command looking the same as the completion
suggestion.

The set_color command in fish supports a fallback color as a second
positional argument, which is documented here:
https://fishshell.com/docs/current/cmds/set_color.html

This commit introduces fallback colors which are recognized by 16 color
terminals. On the Linux console I tested this on, there is no difference
between the bright and non-bright colors, except for black, whose bright
variant is gray.

I did not test this on a 256-color terminal. There it might be
preferable to have no fallback but use interpolation instead.